cleanup uacp conn
